chore: moved frame examples to a examples
